// src/utils/enums/bankTypes.enum.ts
export enum BankTypesEnum {
    default = 0,
    recopays = 1,
    payWing = 2,
    autoPayIn = 3,
    pay365 = 4,
    tappay = 5,
    nexapayUpiCollect = 6,
    firstPe = 7,
    airpay = 8,
    xettle = 9,
    starpaisa = 10,
    finopay = 11,
    finopayStaticQR = 12,
    pay365New = 13,
    OFCPAY = 14,
    theUnionPay = 15,
    heksaPay = 17,
    letspe = 18,
    payTMe = 19,
}

export enum BankTypesEnumValue {
    'Default' = 0,
    'Recopays' = 1,
    'PayWing' = 2,
    'AutoPayIn' = 3,
    'Pay365' = 4,
    'Tappay' = 5,
    'NexapayUpiCollect' = 6,
    'FirstPe' = 7,
    'AirPay' = 8,
    'Xettle' = 9,
    'Starpaisa' = 10,
    'Finopay' = 11,
    'FinopayStaticQR' = 12,
    'Pay365New' = 13,
    'OFCPAY' = 14,
    'TheUnionPay' = 15,
    'HeksaPay' = 17,
    'Letspe' = 18,
    'PayTMe' = 19,
}